Fixed up scheduler and started work on controller monitoring and decoding

+++ RPi/firmware/include/Scheduling.h
@@ -10,7 +10,7 @@ class Scheduling {
 
     public:
         Scheduling() noexcept;
-        Scheduling(int usec) noexcept;
+        Scheduling(double usec) noexcept;
         ~Scheduling(); 
 
         void    RecordTic();
